requirements.txt
requests
wheel
robotframework
robotframework-seleniumlibrary
robotframework-requests
robotframework-xvfb
pyyaml
selenium
selenium-wire
robotframework-imaplibrary2
chromedriver-autoinstaller